Added helm chart source for the operator
This commit includes a new directory, helm, which contains the necessary files for deploying the application to Kubernetes clusters using the Helm package manager.

The chart includes:

- Deployment
- Role and role binding
- Service account
- CRD definition

Resources for the Pulumi Kubernetes Operator, as well as the configurable values file (values.yaml) for specifying the deployment of the Operator.

This change will allow our customers to more easily manage and deploy the Operator on various Kubernetes environments, streamlining the process and reducing the potential for errors.

Signed-off-by: Engin Diri <engin.diri@ediri.de>